requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<0.28
celery>=5.4,<6.0
redis>=5.0,<6.0
psycopg2-binary>=2.9,<3.0
gunicorn>=22.0,<23.0
Pillow>=10.0,<11.0
pytest>=8.2,<9.0
pytest-django>=4.8,<5.0
coverage>=7.5,<8.0
